Add functionality to create story event use case
